+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## Upcomming Version
 - Features:
-
+  - Night order in the side menu for both players and storyteller
 - Bugfixes:
   - [P] Hotkey for adding multiple players
   - Missing message for add multiple players prompt